requirements.txt
Django==1.11.7
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.3
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
markdown2==2.3.5
Pillow==3.1.0
PyJWT==1.5.3
pytz==2017.3
